+++ UIInferno/UICore/_executor/_Markup/Component.ts
@@ -1,1 +1,2 @@
 export { CreateTagVdom } from './Component/CreateTagVdom';
+export { CreateTag } from './Component/CreateTag';
