+++ UIInferno/UICore/_executor/_Markup/Component/CreateTag.ts
@@ -1,0 +1,8 @@
+import { IGeneratorComponent } from 'UICommon/Executor';
+
+
+export class CreateTag implements IGeneratorComponent {
+    create(): void {
+        return;
+    }
+}
